+++ Runtime/EventInterface/InterfaceBindingExtension.cs
@@ -6,5 +6,11 @@ namespace AceLand.EventDriven.EventInterface
     {
         public static void Bind<TInterface>(this object target) =>
             InterfaceBinding.Bind<TInterface>(target);
+        
+        public static void Unbind<TInterface>(this object target) =>
+            InterfaceBinding.Unbind<TInterface>(target);
+        
+        public static bool Implements<TInterface>(this object target) =>
+            InterfaceBinding.Implements<TInterface>(target);
     }
 }